BeepController.cpp: Fixed possible casting bug: (signed) -> (signed long).
Although on my 32-bit Linux, ints and longs are the same length,
they might not be same length elsewhere,
especially on the Arduino the code is for.
So the (signed) cast to inappropriately reduce the resolution.
(signed long) avoids the resolution reduction.
"signed" in (signed long) is unnecessary,
since longs are signed.
I deliberately included "signed" in the (signed long) cast
to emphasis that it is signed.